requirements.txt

fastapi==0.89
uvicorn
mysqlclient
sqlalchemy==1.4
pymysql
alembic
SQLAlchemy-Utils
python-jose[cryptography]
passlib